Move shouldUrlBeOpened helper into separate module
