+++ src/window-creators/helpers/index.js
@@ -1,7 +1,9 @@
 'use strict'
 
 const parseUrl = require('./parse-url')
+const shouldUrlBeOpened = require('./should-url-be-opened')
 
 module.exports = {
-  parseUrl
+  parseUrl,
+  shouldUrlBeOpened
 }
